requirements.txt
# Core
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.1
python-telegram-bot==20.5

# API Clients
python-binance==1.0.19
openai==1.3.0  # GPT API için
ccxt==4.1.13   # Kripto borsaları için
anthropic>=0.5.0  # Claude API için

# Web Scraping ve Arama
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.8.5
selenium==4.11.2
playwright==1.39.0
serpapi-python>=0.1.0
duckduckgo-search>=3.0.0
googlesearch-python>=1.0.0

# Sentiment Analysis
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.0

# Database
sqlalchemy==2.0.20
aiosqlite==0.19.0

# Async Support
asyncio==3.4.3